Updated file structure with Turn Order componennt
